Create pytest-lisa plugin
